requirements.txt
torch
torchvision
transformers
pillow
python-dotenv
flask
huggingface-hub